% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d}
\alias{d}
\title{Example data from quiz 2}
\format{An object of class \code{data.frame} with 50 rows and 2 columns.}
\usage{
d
}
\description{
Example data from quiz2
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q2p1.rda}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
