% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangmengToolsRfunctions.r
\name{func2}
\alias{func2}
\title{Calculate Mean, Variane, SD (again)}
\usage{
func2(x)
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Computes the mean, variance and sd of a vector, but with user checks
}
\examples{
func2(rnorm(10))
}