% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangmengToolsRfunctions.r
\name{func4}
\alias{func4}
\title{Weighted mean, var, sd}
\usage{
func4(d)
}
\arguments{
\item{d}{data.frame}
}
\value{
list
}
\description{
Computes the weighted mean, var, sd
}
\examples{
data(d)
func4(d)
}
