% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangmengToolsRfunctions.r
\name{func5}
\alias{func5}
\title{Weighted mean, var, sd with user checkes}
\usage{
func5(d)
}
\arguments{
\item{d}{data.frame}
}
\value{
list
}
\description{
Computes the weighted mean, var, sd with user checks
}
\examples{
d <- read.table(url("http://www.stat.umn.edu/geyer/3701/data/q1p4.txt"),header = TRUE)
func5(d)
}