% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangmengToolsRfunctions.r
\name{func7}
\alias{func7}
\title{MLE}
\usage{
func7(x, func, interval)
}
\arguments{
\item{x}{vector}

\item{func}{function, e.g., `function(theta, x) dgamma(x, shape = theta, log = TRUE)`}

\item{interval}{vector, i.e., interval for optimize function}
}
\value{
scalar
}
\description{
Computes the liklihood of a given distribution for data x
}
\examples{
func1 = function(theta, x) dgamma(x, shape = theta, log = TRUE)
result7_gamma <- func7(x1,func1,c(0,3))
result7_gamma

}
