% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangmengToolsmyplot.R
\name{plotMyDataA}
\alias{plotMyDataA}
\title{Wrapper function for ggplot2 for data d}
\usage{
plotMyDataA(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
Computes the mean, variance and sd of a vector
}
\examples{
data(d)
plotMyDataA(d)

}