% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangmengToolsmyplot.R
\name{plotMyDataL}
\alias{plotMyDataL}
\title{Wrapper function for ggplot2 for data d}
\usage{
plotMyDataL(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
Computes the mean, variance and sd of a vector
}
\examples{
data(d)
plotMyDataL(d)

}